// src/app/applicant/model/education.model.ts
export interface IEducation {
  grade: any;
  college: any;
  isSpecialTraining: any;
  explainSpecialTraining: any;
  isOtherTraining: any;
  explainOtherTraining: any;
  federalMotoCarrier: any;
  driverBeforeInThisCompany: any;
  dateOfDriving: any;
  reasonUnablePerform: any;
  explainReasonUPerform: any;
  emergencyContact: {}[];
}